trading/requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
requests>=2.31.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
